+++ src/main/java/teamroots/embers/mixin/LateMixin.java
@@ -9,6 +9,7 @@ import java.util.stream.Collectors;
 
 public class LateMixin implements ILateMixinLoader {
     public static final List<String> modMixins = ImmutableList.of(
+            "jaopca",
             "mysticalgears"
     );